+++ fbt_options.py
@@ -20,7 +20,7 @@ DEBUG = 0
 
 # Suffix to add to files when building distribution
 # If OS environment has DIST_SUFFIX set, it will be used instead
-DIST_SUFFIX = "t-dev-009-001"
+DIST_SUFFIX = "t-dev-009-002"
 
 # Post-update slideshow shown by the updater package
 UPDATE_SPLASH = "tumoflip_update"